- Infratstructure & Presentation Layer Modified for sql connection - Entities add configuration - Infratstructure & Presentation Layer Linked to WebUI layer - WebUI layer appsetting modified for sql server connection path - readme file modified for sql server docker query

+++ src/core/Domain/Entities/MovieGenre.cs
@@ -2,7 +2,7 @@
 {
     public class MovieGenre
     {
-        public long MovieId { get; set; }
+        public int MovieId { get; set; }
 
         public int GenreId { get; set; }
